refactor(analysis): extract analysis report model into vailabel-analysis crate
Phase 1 step 8b. Moves the Dataset Intelligence model (AnalysisReport +
health/analytics/quality/image-quality/outlier sub-reports + findings,
ReportSummary, AnalysisConfig, AnalysisJob) into crates/analysis domain, and
the request payloads into contracts. AnalysisReport impls core::Identifiable.
analysis/model.rs is now a glob re-export shim; the engine/service stay in the
binary. Pure serde DTOs, no behavior change.
