+++ apps/studio/src-tauri/Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "crates/runtime",
     "crates/project",
     "crates/dataset",
+    "crates/analysis",
     "crates/annotation",
     "crates/search",
     "crates/training",
@@ -35,6 +36,7 @@ vailabel-core = { path = "crates/core" }
 vailabel-shared = { path = "crates/shared" }
 vailabel-project = { path = "crates/project" }
 vailabel-dataset = { path = "crates/dataset" }
+vailabel-analysis = { path = "crates/analysis" }
 vailabel-plugin = { path = "crates/plugin" }
 vailabel-runtime = { path = "crates/runtime" }
 vailabel-annotation = { path = "crates/annotation" }
@@ -77,6 +79,7 @@ vailabel-core = { workspace = true }
 vailabel-shared = { workspace = true }
 vailabel-project = { workspace = true }
 vailabel-dataset = { workspace = true }
+vailabel-analysis = { workspace = true }
 diesel = { version = "2", features = ["sqlite"] }
 libsqlite3-sys = { version = "0.28", features = ["bundled"] }
 reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls", "json"] }
