+++ apps/studio/src-tauri/crates/analysis/src/contracts/mod.rs
@@ -1,0 +1,31 @@
+//! Inbound request payloads for the analysis commands.
+
+use serde::Deserialize;
+
+use crate::domain::AnalysisConfig;
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct AnalysisRequest {
+    pub project_id: String,
+    #[serde(default)]
+    pub config: AnalysisConfig,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ReportIdPayload {
+    pub id: String,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ProjectIdPayload {
+    pub project_id: String,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct JobIdPayload {
+    pub job_id: String,
+}
